CRIU j9time_nano_time skips the elapsed time between Checkpoint/Restore
Added a field checkpointRestoreTimeDelta within J9PortLibrary, which is
to be set the time delta between Checkpoint and Restore of
j9time_nano_time() return values;
Updated trace points to record the actual j9time_nano_time() values;
Add a test to verify the elapsed time.
