+++ test/functional/cmdLineTests/criu/criuScript.sh
@@ -23,15 +23,15 @@
 #
 
 echo "start running script";
-$2 -XX:+EnableCRIUSupport $3 -cp "$1/criu.jar" CRIUSimpleTest $4 >testOutput 2>&1;
+$2 -XX:+EnableCRIUSupport $3 -cp "$1/criu.jar" $4 $5 >testOutput 2>&1;
 sleep 2;
 criu restore -D ./cpData --shell-job;
-if [ "$4" == "TwoCheckpoints" ] || [ "$4" == "ThreeCheckpoints" ]
+if [ "$5" == "TwoCheckpoints" ] || [ "$5" == "ThreeCheckpoints" ]
 then
 sleep 2;
 criu restore -D ./cpData --shell-job;
 fi
-if [ "$4" == "ThreeCheckpoints" ]
+if [ "$5" == "ThreeCheckpoints" ]
 then
 sleep 2;
 criu restore -D ./cpData --shell-job;
