+++ test/functional/cmdLineTests/criu/src/org/openj9/criu/CRIURandomTest.java
@@ -19,6 +19,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
+package org.openj9.criu;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
